+++ server/src/main/java/org/allaymc/server/container/processor/ConsumeActionProcessor.java
@@ -69,7 +69,7 @@ public class ConsumeActionProcessor implements ContainerActionProcessor<ConsumeA
                                                 item.getCount(),
                                                 item.getStackNetworkId(),
                                                 item.getCustomName(),
-                                                item.getDurability(),
+                                                item.getDamage(),
                                                 ""
                                         )
                                 ),