+++ server/src/main/java/org/allaymc/server/container/processor/DestroyActionProcessor.java
@@ -62,7 +62,7 @@ public class DestroyActionProcessor implements ContainerActionProcessor<DestroyA
                                                 item.getCount(),
                                                 item.getStackNetworkId(),
                                                 item.getCustomName(),
-                                                item.getDurability(),
+                                                item.getDamage(),
                                                 ""
                                         )
                                 ),